+++ packages/maps/src/lib/index.js
@@ -12,29 +12,38 @@ export { default as MapControlRefresh } from './mapControlRefresh/MapControlRefr
 export { default as MapControlZoom } from './mapControlZoom/MapControlZoom.svelte';
 
 // Location Search
-export * from './mapControlLocationSearch/MapGeocoderAdapterMapBox';
 export { default as MapControlGeocoder } from './mapControlLocationSearch/MapControlGeocoder.svelte';
 export { default as MapControlGeolocator } from './mapControlLocationSearch/MapControlGeolocator.svelte';
 export { default as MapControlLocationSearch } from './mapControlLocationSearch/MapControlLocationSearch.svelte';
+export * from './mapControlLocationSearch/MapGeocoderAdapterMapBox';
 
 // Borough Selector
 export { default as MapControlBorough } from './mapControlBorough/MapControlBorough.svelte';
+
+// Legend
+export { default as MapLegend } from './mapLegend/MapLegend.svelte';
 
 // Events
 export { default as MapCursorEvent } from './mapCursorEvent/MapCursorEvent.svelte';
 
 // Popups
-export { default as MapMarker } from './mapMarker/MapMarker.svelte';
 export { default as MapMarkerContainer } from './mapMarker/elements/mapMarkerContainer/MapMarkerContainer.svelte';
 export { default as MapMarkerFlyToFeature } from './mapMarker/elements/mapMarkerFlyToFeature/MapMarkerFlyToFeature.svelte';
 export { default as MapMarkerPlacement } from './mapMarker/elements/mapMarkerPlacement/MapMarkerPlacement.svelte';
 export { default as MapMarkerStyledContainer } from './mapMarker/elements/mapMarkerStyledContainer/MapMarkerStyledContainer.svelte';
+export { default as MapMarker } from './mapMarker/MapMarker.svelte';
+
+export { default as MapDeckPopovers } from './mapDeckPopovers/MapDeckPopovers.svelte';
+export { onClickPopoverHandler, clickedLayer, clickedFeature } from './mapDeckPopovers/stores';
+
+export { default as MapDeckTooltips } from './mapDeckTooltips/MapDeckTooltips.svelte';
+export { onMouseOverTooltipHandler, mousedOverObject } from './mapDeckTooltips/stores';
 
 export { default as MapPopover } from './mapPopover/MapPopover.svelte';
 
 // Layers
-export { default as MapLayerSource } from './mapLayerSource/MapLayerSource.svelte';
 export { default as GeoJSONMapLayerSource } from './mapLayerSource/adaptations/geojsonMapLayerSource/GeoJSONMapLayerSource.svelte';
+export { default as MapLayerSource } from './mapLayerSource/MapLayerSource.svelte';
 export { default as MapLayerView } from './mapLayerView/MapLayerView.svelte';
 
 // Context Layers
@@ -47,7 +56,7 @@ export { default as MapDeckOverlay } from './mapDeckOverlay/MapDeckOverlay.svelt
 export * from './themes/animations';
 export * from './themes/bounds';
 
-export * as theme_os_dark_grey_muted_buildings from './themes/os_dark_grey_muted_buildings.json';
 export * as theme_os_dark from './themes/os_dark.json';
+export * as theme_os_dark_grey_muted_buildings from './themes/os_dark_grey_muted_buildings.json';
 export * as theme_os_greyscale from './themes/os_greyscale.json';
 export * as theme_os_light_vts from './themes/os_light_vts.json';