+++ packages/tables/src/lib/core/lib/types.ts
@@ -81,6 +81,11 @@ export type ColSpec = {
 		renderer?: string | SvelteComponent;
 		[other: string]: any;
 	};
+
+	header?: {
+		renderer?: string | SvelteComponent;
+		[other: string]: any;
+	};
 };
 
 // TODO: check