+++ packages/tables/src/lib/core/renderers/Header.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
-	import { Tooltip, classNames } from '@ldn-viz/ui';
+	import { classNames, Overlay } from '@ldn-viz/ui';
 	import { BarsArrowDown, ChevronUpDown } from '@steeze-ui/heroicons';
 	import { Icon } from '@steeze-ui/svelte-icon';
+	import type { ComponentType } from 'svelte';
 
 	/**
 	 * Text of label/column heading.
@@ -29,31 +30,44 @@
 	export let toggle: (ev: Event) => void = () => {};
 
 	/**
-	 * Text to be displayed in tooltip after user hovers on info icon after column heading.
+	 * Text to be displayed in tooltip/popover/modal after user hovers or clicks on info icon after column heading.
 	 */
 	export let hintText = '';
 
-	export let alignHeader: 'left' | 'right' | 'center' | undefined;
+	/**
+	 * A Svelte component to be displayed instead of hintText.
+	 */
+	export let hintComponent: undefined | ComponentType = undefined;
+
+	/**
+	 * The type of overlay in which the hint will be displayed.
+	 */
+	export let hintType: 'tooltip' | 'popover' | 'modal' = 'tooltip';
+
+	/**
+	 * The text color.
+	 */
+	export let color = '';
+
+	export let alignHeader: 'left' | 'right' | 'center' | undefined = 'left';
 	const alignmentClasses = {
 		left: 'justify-start',
 		right: 'justify-end',
 		center: 'justify-center'
 	};
-	$: alignmentClass = alignmentClasses[alignHeader ?? 'left'];
+	$: alignmentClass = alignmentClasses[alignHeader];
 
 	// This suppresses warnings due to the RowRenderer providing props that aren't used.
 	// eslint-disable-next-line @typescript-eslint/no-unused-expressions
 	$$restProps;
 </script>
 
-<div
-	on:click={toggle}
-	on:keypress={toggle}
-	role="cell"
-	tabindex={0}
-	class="font-bold py-0.5 w-full h-full"
->
-	<div class={classNames('flex items-center select-none', alignmentClass)}>
+<div class="font-semibold py-0.5 w-full h-full" style:color>
+	<svelte:element
+		this={allowSorting ? 'button' : 'div'}
+		class={classNames('flex items-center select-none', alignmentClass)}
+		on:click={allowSorting ? toggle : undefined}
+	>
 		{#if superscriptText}
 			<div class="text-left">
 				<span class="font-normal text-xs">{superscriptText}</span><br />
@@ -61,8 +75,15 @@
 			</div>
 		{:else}
 			{label}
-			{#if hintText}
-				<Tooltip hintLabel="" hintSize="sm">{hintText}</Tooltip>
+
+			{#if hintText || hintComponent}
+				<Overlay hintLabel="" overlayType={hintType} modalTitle={label}>
+					{#if hintComponent}
+						<svelte:component this={hintComponent} />
+					{:else}
+						{hintText}
+					{/if}
+				</Overlay>
 			{/if}
 		{/if}
 
@@ -74,7 +95,7 @@
 				aria-hidden="true"
 			/>
 		{/if}
-	</div>
+	</svelte:element>
 </div>
 
 <style>