+++ packages/ui/src/lib/layoutExamples/Dashboards/SimpleDashboard.svelte
@@ -17,12 +17,11 @@
 	<svelte:fragment slot="main">
 		<div class="bg-color-canvas-background-1 h-full p-6">
 			<div class="mb-4">
-				<h1 class="headline responsive">This is the main content area</h1>
-				<h2 class="subhead responsive font-normal">
+				<h1 class="title-lg responsive">This is the main content area</h1>
+				<h2 class="subtitle responsive font-normal">
 					Maecenas ut libero vel nibh maximus feugiat non sed tortor.
 				</h2>
 			</div>
-			<div class="bg-color-static-white">HELLO</div>
 		</div>
 	</svelte:fragment>
 	<Sidebar slot="sidebar">