+++ packages/ui/src/lib/sidebar/Sidebar.svelte
@@ -13,6 +13,7 @@
 	import { type Writable } from 'svelte/store';
 	import { slide } from 'svelte/transition';
 	import { classNames } from '../utils/classNames';
+	import { randomId } from '../utils/randomId';
 	import TabPanel from './../tabs/TabPanel.svelte';
 	import type { Tab } from './../tabs/types';
 	import SidebarTabList from './elements/sidebarTabs/SidebarTabList.svelte';
@@ -64,9 +65,19 @@
 	$: component = tabs.find((tab) => tab.id === selectedValue)?.content;
 
 	/**
+	 * Aria label to describe purpose of sidebar
+	 */
+	export let sidebarAriaLabel: string = 'Sidebar with information and controls';
+
+	/**
 	 * Aria label applied to tabs list
 	 */
-	export let ariaLabel: string = 'Switch sidebar panel';
+	export let tabsAriaLabel: string = 'Switch sidebar panel';
+
+	/**
+	 * Randomly generated id for sidebar container. Used by `SidebarToggle` to tell screen readers what the toggle controls.
+	 */
+	export let sidebarId: string = randomId();
 
 	const sidebarPlacementFromContext = getContext<Writable<PlacementType>>('sidebarPlacement');
 	const sidebarIsOpen = getContext<Writable<boolean>>('sidebarIsOpen');
@@ -92,16 +103,18 @@
 	{#if tabs.length}
 		<div class={classNames('absolute bg-color-container-level-0', tabPlacementClasses)}>
 			<!-- A `<SidebarTabList>`, if the sidebar has tabs-->
-			<SidebarTabList {tabs} {ariaLabel} bind:selectedValue />
+			<SidebarTabList {tabs} {placement} ariaLabel={tabsAriaLabel} bind:selectedValue />
 		</div>
 	{:else if $sidebarAlwaysOpen !== 'true'}
 		<div class={classNames('absolute', togglePlacementClasses)}>
-			<SidebarToggle />
+			<SidebarToggle {sidebarId} />
 		</div>
 	{/if}
 
 	{#if $sidebarIsOpen}
-		<div
+		<aside
+			id={sidebarId}
+			aria-label={sidebarAriaLabel}
 			class={classNames('flex', heightClasses)}
 			transition:slide={{ duration: 300, axis: transitionAxis[placement] }}
 		>
@@ -160,6 +173,6 @@
 					</div>
 				{/if}
 			</div>
-		</div>
+		</aside>
 	{/if}
 </div>